FInal Capstone Project of Android Developer Course at Udacity
News App is an app which display a list of selected topic news and the api for this app has been taken from​ ​newsapi.org
In this app, you can see the news headlines, details and you can bookmark your news. The widget for this app displays the bookmarked news which when pressed will direct you to the particular news details in the app.